requirements.txt
# 核心框架和AI工具
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.0.1

# AI服务和API
openai>=1.50.0
dashscope>=1.20.0

# 向量数据库和嵌入模型
langchain-huggingface>=0.0.3
sentence-transformers>=2.3.0
chromadb>=0.4.0
sentence-transformers>=2.3.0
text2vec>=1.3.0

# Web框架和服务器
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
aiohttp>=3.9.0
httpx>=0.27.0

# 数据库和ORM
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# 数据处理和科学计算
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.3.0

# 文档处理
docx2txt>=0.8
unstructured>=0.10.0
langchain-community>=0.2.0
python-docx>=1.0.0
chardet>=5.0.0

# 配置和环境管理
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# 日期和日历
lunarcalendar>=0.0.9
chinesecalendar>=1.0.0

# 监控和日志
prometheus-client>=0.20.0
structlog>=24.1.0

# 开发工具
black>=24.0.0
flake8>=7.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
colorama>=0.4.6
tqdm>=4.65.0